異常値調整のロジック変更

+++ Assets/Scripts/MainScene.cs
@@ -30,7 +30,7 @@ public class MainScene : MonoBehaviour
         int fpsIndex = DefaultValue.FPS_INDEX;
         int adjustAbnormalPosition = DefaultValue.ADJUST_ABNORMAL_POSITION;
         int smooth = DefaultValue.SMOOTH;
-        int thresholdMovePos = DefaultValue.THRESHOLD_MOVE_POS;
+        float thresholdMovePos = DefaultValue.THRESHOLD_MOVE_POS;
 
         if (PlayerPrefs.HasKey(PlayerPrefsKey.IP))
         {
@@ -54,7 +54,7 @@ public class MainScene : MonoBehaviour
         }
         if (PlayerPrefs.HasKey(PlayerPrefsKey.THRESHOLD_MOVE_POS))
         {
-            thresholdMovePos = PlayerPrefs.GetInt(PlayerPrefsKey.THRESHOLD_MOVE_POS);
+            thresholdMovePos = PlayerPrefs.GetFloat(PlayerPrefsKey.THRESHOLD_MOVE_POS);
         }
 
         uClient.enabled = false;
